pyproject.toml
[project]
name = "nofluffjobs-llm-matcher"
version = "0.1.0"
description = "A tool to scrape NoFluffJobs using Playwright and evaluate job fit with Qwen 3.5 LLM."
readme = "README.md"
requires-python = ">=3.13"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "playwright>=1.42.0",
    "playwright-stealth>=2.0.0",
    "fake-useragent>=1.5",
    "vllm; sys_platform != 'win32'",
    "rich>=13.7.1",
    "python-dotenv>=1.0.1",
    "tenacity>=8.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = [
    "ruff>=0.3.4",
    "mypy>=1.9.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = []

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src/nofluffjobs_llm_matcher"]

[project.scripts]
nfj-matcher = "nofluffjobs_llm_matcher.main:cli_entry"
